+++ internal/provider/permissions_graph_resource_test.go
@@ -4,18 +4,11 @@ import (
 	"fmt"
 	"testing"
 
+	"github.com/flovouin/terraform-provider-metabase/metabase"
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 )
 
-func testAccPermissionsGraphResource(setData bool) string {
-	data := ""
-	if setData {
-		data = `data = {
-        native = "write"
-        schemas = "all"
-      }`
-	}
-
+func testAccPermissionsGraphResource(createQueries string) string {
 	return fmt.Sprintf(`
 import {
   to = metabase_permissions_graph.graph
@@ -30,15 +23,15 @@ resource "metabase_permissions_graph" "graph" {
       group    = 1
       database = 1
       download = {
-        native  = "full"
         schemas = "full"
       }
-			%s
+      view_data = "unrestricted"
+      create_queries = "%s"
     },
   ]
 }
 	`,
-		data,
+		createQueries,
 	)
 }
 
@@ -47,14 +40,14 @@ func TestAccPermissionsGraphResource(t *testing.T) {
 		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
 		Steps: []resource.TestStep{
 			{
-				Config: providerApiKeyConfig + testAccPermissionsGraphResource(true),
+				Config: providerApiKeyConfig + testAccPermissionsGraphResource(string(metabase.PermissionsGraphDatabasePermissionsCreateQueriesQueryBuilderAndNative)),
 				Check: resource.ComposeAggregateTestCheckFunc(
 					resource.TestCheckResourceAttr("metabase_permissions_graph.graph", "advanced_permissions", "false"),
 					resource.TestCheckResourceAttrSet("metabase_permissions_graph.graph", "revision"),
 				),
 			},
 			{
-				Config: providerApiKeyConfig + testAccPermissionsGraphResource(false),
+				Config: providerApiKeyConfig + testAccPermissionsGraphResource(string(metabase.PermissionsGraphDatabasePermissionsCreateQueriesNo)),
 				Check: resource.ComposeAggregateTestCheckFunc(
 					resource.TestCheckResourceAttr("metabase_permissions_graph.graph", "advanced_permissions", "false"),
 					resource.TestCheckResourceAttrSet("metabase_permissions_graph.graph", "revision"),